Create README.md
This app was created to honor my grandmother who died several years ago. My goal is to create a useable recipe app with my grandmother's recipes and many other recipes. I would like to place this app in the app store so my extended family can have access to it. I would like users to be able to modify and add recipes.